pyproject.toml
[project]
name = "backtest-engine"
version = "0.1.0"
description = "Vectorized backtesting engine for daily-bar equity strategies"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "yfinance>=0.2",
    "pyarrow>=16.0",
    "matplotlib>=3.8",
    "vectorbt>=0.26",
    "alpaca-py>=0.8",
    "python-dotenv>=1.0",
    "streamlit>=1.35",
    "plotly>=5.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "pandas-stubs>=2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"
